requirements.txt
numba
colorama
tensorflow-macos